backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=1.4.0
psycopg2-binary  # Or your specific DB driver if not using default SQLite
pydantic>=1.9.0
python-dotenv>=0.19.0
sentence-transformers>=2.2.0
numpy>=1.18.0
typing-extensions>=4.0.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
requests>=2.28.0
openai>=1.4.0
# Uncomment for production use
# redis>=4.3.0 